Exported the part of the code into a library as a test
