+++ DiceChance.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 15.0.28307.489
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DiceChance", "DiceChance\DiceChance.vcxproj", "{030CD7DA-9BED-49DE-AA21-7A3A4CE011B4}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DiceChanceLibrary", "DiceChanceLibrary\DiceChanceLibrary.vcxproj", "{82D1B1D2-4863-46AD-A760-D1E19AE1B979}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -21,6 +23,14 @@ Global
 		{030CD7DA-9BED-49DE-AA21-7A3A4CE011B4}.Release|x64.Build.0 = Release|x64
 		{030CD7DA-9BED-49DE-AA21-7A3A4CE011B4}.Release|x86.ActiveCfg = Release|Win32
 		{030CD7DA-9BED-49DE-AA21-7A3A4CE011B4}.Release|x86.Build.0 = Release|Win32
+		{82D1B1D2-4863-46AD-A760-D1E19AE1B979}.Debug|x64.ActiveCfg = Debug|x64
+		{82D1B1D2-4863-46AD-A760-D1E19AE1B979}.Debug|x64.Build.0 = Debug|x64
+		{82D1B1D2-4863-46AD-A760-D1E19AE1B979}.Debug|x86.ActiveCfg = Debug|Win32
+		{82D1B1D2-4863-46AD-A760-D1E19AE1B979}.Debug|x86.Build.0 = Debug|Win32
+		{82D1B1D2-4863-46AD-A760-D1E19AE1B979}.Release|x64.ActiveCfg = Release|x64
+		{82D1B1D2-4863-46AD-A760-D1E19AE1B979}.Release|x64.Build.0 = Release|x64
+		{82D1B1D2-4863-46AD-A760-D1E19AE1B979}.Release|x86.ActiveCfg = Release|Win32
+		{82D1B1D2-4863-46AD-A760-D1E19AE1B979}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
