+++ DiceChance/Source.cpp
@@ -1,104 +1,46 @@
 #include <iostream>
 #include <iomanip>
 #include <string>
-
+#include "DiceChance.h"
 using namespace std;
-typedef long long llong;
-
-// Возможные виды вывода шанса
-enum VsType{
-	NoVs, GreaterThan, LessThan, Equals
-};
-
-// Возвращает массив, в котором записано количество возможных совпадений.
-int* GetValueSpread(int cubes, int sides, int &N) {
-	if (cubes < 1 || sides < 1) return NULL;
-
-	N = sides * cubes + 1;
-	int **a = new int*[cubes + 1];
-	for (int i = 0; i <= cubes; i++)
-		a[i] = new int[sides * cubes + 1];
-
-	for (int i = 0; i <= cubes; i++)
-		for (int j = 0; j <= sides * cubes; j++)
-			a[i][j] = 0;
-	for (int j = 1; j <= sides; j++) a[1][j] = 1;
-
-	for (int i = 2; i <= cubes; i++) {
-		for (int j = i; j <= i * sides; j++)
-			a[i][j] = a[i][j - 1] + a[i - 1][j - 1] - a[i - 1][i * (sides + 1) - j];
+using namespace Dice;
+// Returns the number of digits in a number
+int GetNumberOfDigits(int a) {
+	int k = 0;
+	while (a > 0) {
+		a /= 10; k++;
 	}
-
-	int* result = a[cubes];
-	for (int i = 0; i < cubes; i++)
-		delete[] a[i];
-	delete[] a;
-	for (int i = cubes; i < N; i++) {
-		if (result[i] < 0) throw overflow_error("Too many possible values");
-	}
-	return result;
-}
-
-// Возвращает общее количество совпадений
-double GetTotal(int *a, int N) {
-	double total = 0;
-	for (int i = 0; i < N; i++) total += a[i];
-	return total;
-}
-
-// Возвращает массив шансов по заданному массиву совпадений
-double* GetChanceArray(int *a, int N) {
-	double total = GetTotal(a, N);
-	double* chance = new double[N];
-	for (int i = 0; i < N; i++)
-		chance[i] = a[i] / total;
-	return chance;
-}
-
-// Возвращает шанс по заданному массиву совпадений и типу
-double GetChance(int *a, int N, int vsNum, VsType type) {
-	double total = GetTotal(a, N);
-	double chance = 0;
-	int k = 1;
-	while (k <= vsNum)
-		chance += a[k++] / total;
-	switch (type) {
-	case Equals:
-		return a[vsNum] / total;
-	case LessThan:
-		return chance;
-	case GreaterThan:
-		chance = 1 - chance;
-		chance += a[vsNum] / total;
-		return chance;
-	}
-	return 0;
-}
-
-// Складывает два массива совпадений и возвращает результат
-int* AddArray (int *a, int Na, int *b, int Nb, int& N) {
-	N = Na + Nb - 1;
-	int *c = new int[N];
-	for (int i = 0; i < N; i++)
-		c[i] = 0;
-	for (int i = 1; i < Na; i++)
-		for (int j = 1; j < Nb; j++)
-			c[i + j] += a[i] * b[j];
-	return c;
+	return k;
 }
 
 // Удаляет из начала и конца строки знаки пробелов и табуляции
-void CleanString(string &s) {
+void CleanString(std::string &s) {
 	int n = 0;
 	while (n < s.length() && (s[n] == ' ' || s[n] == '\t')) n++;
 	s.erase(0, n);
 	n = s.length() - 1;
 	while (n >= 0 && (s[n] == ' ' || s[n] == '\t')) n--;
-	s.erase(n+1);
+	s.erase(n + 1);
+}
+// Возвращает stoi(s) или defaultNum
+int SafeStoi(std::string &s, int defaultNum) {
+	int result;
+	try {
+		result = stoi(s);
+	}
+	catch (const std::exception e) {
+		result = defaultNum;
+	}
+	return result;
+}
+
+// Является ли данный символ цифрой
+bool isNum(char a) {
+	return a >= '0' && a <= '9';
 }
 
 // Находит и возвращает тип вывода шанса, если таковой есть
-VsType GetType(string &s, int &vsNum) {
+Dice::VsType GetType(string &s, int &vsNum) {
 	VsType vsType = NoVs;
 	vsNum = s.find('=');
 	if (vsNum != -1) vsType = Equals;
@@ -112,23 +54,10 @@ VsType GetType(string &s, int &vsNum) {
 	}
 	if (vsType != NoVs) {
 		string vsStr = s.substr(vsNum + 1);
-		try { vsNum = stoi(vsStr); }
-		catch (const exception &e) { vsNum = -1; }
+		vsNum = SafeStoi(vsStr, -1);
 	}
 	else vsNum = 0;
 	return vsType;
-}
-
-// Возвращает stoi(s) или defaultNum
-int SafeStoi(string &s, int defaultNum) {
-	int result;
-	try {
-		result = stoi(s);
-	}
-	catch (const exception e) {
-		result = defaultNum;
-	}
-	return result;
 }
 
 // Возвращает количество кубиков в выражении или -1, если их нет
@@ -136,11 +65,6 @@ int GetCubes(string &s, int dPos) {
 	string cubeStr = s.substr(0, dPos);
 	CleanString(cubeStr);
 	return SafeStoi(cubeStr, -1);
-}
-
-// Является ли данный символ цифрой
-bool isNum(char a) {
-	return a >= '0' && a <= '9';
 }
 
 // Возвращает количество сторон в выражении или -1, если их нет
@@ -165,35 +89,6 @@ int GetPlus(string &s) {
 	return SafeStoi(opStr, 0);
 }
 
-// Смещает массив на plus. При смещении элемент с индексом 0 хранит в себе все элементы, индекс которых меньше 0
-void AdjustArray(int *&a, int &N, int plus) {
-	if (plus == 0) return;
-	int *b, Nb = N;
-	if (plus > 0) 
-		Nb += plus;
-	b = new int[Nb];
-	for (int i = 0; i < Nb; i++)
-		b[i] = 0;
-	for (int i = 0; i < N; i++) {
-		int id = i + plus;
-		if (id < 0) id = 0;
-		b[id] += a[i];
-	}
-	N = Nb;
-	int *t = a;
-	a = b;
-	delete[] t;
-}
-
-// Возвращает количество цифр в числе
-int GetNumberOfDigits(int a) {
-	int k = 0;
-	while (a > 0) {
-		a /= 10; k++;
-	}
-	return k;
-}
-
 int main() {
 	int cubes, sides, plus, vsNum;
 	string s;
@@ -213,14 +108,7 @@ int main() {
 	} while (!sucess);
 	int N;
 	int *a;
-	try {
-		a = GetValueSpread(cubes, sides, N);
-	}
-	catch (const overflow_error e) {
-		cout << e.what();
-		cin.get();
-		return 0;
-	}
+	a = GetValueSpread(cubes, sides, N);
 	AdjustArray(a, N, plus);
 	double *chances = GetChanceArray(a, N);
 	double chance = GetChance(a, N, vsNum, vsType);
