+++ StaticLib1/stdafx.cpp
@@ -1,0 +1,1 @@
+﻿#include "stdafx.h"
