+++ StaticLib1/stdafx.h
@@ -1,0 +1,14 @@
+﻿// stdafx.h: включаемый файл для стандартных системных включаемых файлов
+// или включаемых файлов для конкретного проекта, которые часто используются, но
+// нечасто изменяются
+//
+
+#pragma once
+
+#include "targetver.h"
+
+#define WIN32_LEAN_AND_MEAN             // Исключите редко используемые компоненты из заголовков Windows
+
+
+
+// установите здесь ссылки на дополнительные заголовки, требующиеся для программы
